don't delete staged dir on submission error
In machete we delete the job on submission error because we stage the job during submission
so it makes sense in this way to clean yourself up.

But in job constructor we changed the stage method to
not stage if it was already staged, because we actually stage earlier.

So we override submit so that we don't delete if an error occurs.