hide columns and fix view for Firefox & Safari
By adding a margin of 5px or larger, the view is fixed in Firefox and Safari
For some reason the entire ul block was smushed to the far right as though.
It were an inline-block and there was no newline between the previous fm-header.

The rest was as before: display none on mode and owner, and then make the
width of the name column wider.