Use viewport for sizing of the file manager div
the fm div wraps everything, so we size this to height of 100vh and use
padding with a box-sizing: border-box; so that the padding is included
in the height. The result is there is spacing between the edge of the
viewport and the contents and also the contents stay within the
viewport, resizing as appropriate using flexbox

this only works in IE10+ and appears to work in Safari 8, Chrome 50, and
Firefox  34.