add hide/show class for mode & owner
The hide-owner-mode css class on the body (or above the panel) will hide these rows.
By default these are supposed to be hidden so I updated the body tag to
have this class. In the future this should be toggle-able via a
checkbox.

This is using flexbox completely now - no more arbitrary percent widths
on columns.