Update joyride text on main page, change order
Updated joyride text to emphasise the three steps:

1. Create
2. Edit
3. Submit

I also changed the order so the Create Template help text is displayed
at the end.

Finally, commented out the "My Templates" help for now as the tooltip is
visibly off and we might want to remove the "My Templates" page in the
future in favor of just having this be merged with the "Create Job" page
where the templates are.