add user icon to fix font awesome issue
Chrome (and maybe other browsers) loads font awesome after the first
character is rendered using that font. OnDemand only displays icons in
dropdown menus, which means often when you have a dashboard open and then
you are logged out because of a session expiration you end up seeing
boxes where there should be icons.

Solution is to add at least one font awesome icon on the page that is
visible on load. We add the user icon next to the logged in user so we can
solve this issue AND this gives us the ability to hide "logged in as" text
when the browser window gets too small.