added Workflow.active and Workflow.update_status!
By using the Workflow.active scope, we can remove from the WorkflowsController
working with Job directly when updating the status of all jobs.

The scope + preload(:jobs) ensures only 1 query is used to get all the data, but
we can now iterate over the active workflows and call update_status on these.

This is a step towards replacing the osc_machete_rails and osc-machete code with
the newer ood_job and ood_job_rails solutions.