Avoid using class methods in configuration
Using Configuration as a set of class methods makes it difficult
to write tests and of course has other problems.

Following the first suggestion in the blog article

https://8thlight.com/blog/josh-cheek/2012/10/20/implementing-and-testing-the-singleton-pattern-in-ruby.html

after defining the Configuration class with instance methods and variables
for the config, we then store a global instance of Configuration to a new constant:

AppConfig = Configuration.new

The rest of the app uses AppConfig to access the values for this configuration.

Now AppConfig is just an instance of Configuration, so I can create new
Configuration instances in tests, or I can marshal the AppConfig instance
from another instance of my app.