Fix UsrRouter#caption error f non-user owned app
For usr apps both the caption and the default category are based on the UsrRouter#owner_title
which was looking up the full name of the username using gecos field. The problem is that now
the owner of usr apps can be a non-user i.e. a group, or really any directory under
/var/www/ood/apps/usr/ that is setup correctly.