fix inconsistency with portal & title env vars
ONDEMAND_TITLE and ONDEMAND_PORTAL were added previously when
the apps were using OOD_PORTAL and OOD_DASHBOARD_TITLE

So the logic was

1. set the new env vars to default values unless ondemand_portal
   and ondemand_title specified in nginx_stage.yml
2. set OOD_PORTAL and OOD_DASHBOARD_TITLE if ondemand_portal
   and ondemand_title specified in nginx_stage.yml

The problem is that in the first case setting ONDEMAND_TITLE also stripped
whitespace, and setting ONDEMAND_PORTAL also stripped whitespace and made
the string lowercase and substituded remaining whitespace for underlines.

OOD_PORTAL and OOD_DASHBOARD_TITLE got the value in nginx_stage.yml as is.

This commit fixes this inconsistancy while retaining the original logic above.