nginx_stage: generate secret key and set to env
Generate a unique secrete key for each user and
set it to SECRET_KEY_BASE env var.

Generate when fetching NginxStage.nginx_env
Cache the secret in a file next to the pun config if the cache ! exist.
If the cache file exists, the secret is read from the file instead of being generated