Improve time reporting
Overrides the Rails distance_of_time_in_words with the dotiw gem. This gives us the ability to report exact time diffs in words accumulating on the type we want. So two days of compute renders as 48 hours. 3600 seconds renders as 1 hour instead of about 1 hour. Due to the imprecision in our set up we are not reporting on seconds. dotiw provides backwards compatible functionality using the :vague keyword if we need that elsewhere in the application.