Add optional whitelist for file editor
Utilize the environment variable WHITELIST_PATH to define an optional set of paths that are the only paths that can be viewed/edited. An empty whitelist permits viewing/editting of any path that the user would normally have permissions for. If the whitelist is defined then, any requested path that is not in rooted in a member of the whitelist returns an error message and a 403, before checking if the requested path actually exists so as not to leak information about the filesystem.