nginx_stage: set OOD_DEV_APPS_ROOT
This is so app development mode can be enabled by default if the symlink is created to the user's dev directory.

Currently the dashboard needs to provide this custom code in the initializer:

    Configuration.app_development_enabled = File.directory? "/var/www/ood/apps/dev/#{OodSupport::Process.user.name}/gateway"

Now the dashboard can be updated to use OOD_DEV_APPS_ROOT for the path to the dev apps root for the given user.