Handle edge cases which could crash the app, and improve handling of limits
At OSC usage of the scratch file system is not limited. In the future it could be limited for some or all users, and so it should not be ignored. Given that there there may be file systems or users for which a resource is unlimited Quota#valid? has been changed to Quota#limited?. To ensure that the quota files are kept sane a warning has been added which will trigger if the value for limit is NOT one of:

- 0
- A string which #to_i turns into a positive integer
- nil
- the string which when downcased becomes 'unlimited'

In cases where an invalid limit is detected the limit is set to 0 to avoid a crash, and a warning message is logged