Add improved logic for handling heterogenous clusters
Add support for showing or hiding job array controls, depending on whether or not any clusters support job arrays. For example imagine a system where there are two clusters; one for which there is job array support, and one where there is not.

On the job details edit page the job array input will be shown because at least one cluster is supported. If the user selects the non-supported cluster then the control will disable and blank. Since not job arrays are not supported on all clusters the the help message for the job array request will notify the user that job arrays are optional and not all clusters are supported.