Adds the ability to SSH to multiple hosts from the file explorer
If the environment variable `OOD_SSH_HOSTS` is set then a dropdown button will appear in place of the regular single click button. `OOD_SSH_HOSTS` is expected to be a comma delimited list of colon delimited pairs. The first half of the pair needs to be the URL to connect to the shell for and must contain the full connection URL (/pun/sys/shell/ssh/...), and the second should be a human-friendly label for the cluster.

For example at OSC we would use '/pun/sys/shell/ssh/owens.osc.edu:Owens,/pun/sys/shell/ssh/ruby.osc.edu:Ruby,/pun/sys/shell/ssh/pitzer.osc.edu:Pitzer'.

Note that there is no support at this time for filtering clusters, which is not ideal in cases where not all users have access to a given cluster.

Other decisions of note:

- The rendering engine has been partially changed to Handlebars instead of Rendy to support conditionals and loops
- Bootstrap CSS and JS have been included with the JS coming from the same CDN that we are using for jquery