Improve default dev mode enabling
By default dev enabling in 1.4+ works like shared apps enabling,
where a symlink to the dev directory is created, such as
/var/www/ood/apps/dev/efranz/gatway => ~efranz/ondemand/dev

Before, we only enabled if this directory existed; so after the admin
adds the symlink, the user must create the dev directory.

With this change we enable dev mode if the symlink exists, even if the
corresponding directory has not yet been created. Since the dashboard
will create the directory if it doesn't yet exist when navigating to it,
this removes an extra step to enabling dev mode in the dashboard.