Add configurable logo height
By setting OOD_DASHBOARD_LOGO_HEIGHT to a value such as `100px` you can set a fixed height on the logo. This enables the use of SVG images for the logo without the logo expanding the full width of the page.

Also added same margin bottom to the logo as is on the default logo.