batch_connect: specify cluster in submit.yml #553
Fix for #529. Users can now specify the cluster in the form.yml.erb or the submit.yml.erb files. 
If not defined in either, the dashboard now throws a translatable error message. 

cluster is expected to be a string on the top level yaml of the submit.yml file.  Something like this:

    <%-
      cluster = 'owens'
    -%>
    ---
    cluster: "<%= cluster %>"
    batch_connect:
      template: "basic"