don't always use cached clusters  (#748)
Don't use cached clusters when the cached cluster value conflicts with the current form

As the comment in delete_cached_cluster? says, it's possible to (a) have a cached
cluster value that no longer exists OR the app has simply reconfigured a different
cluster to use.

This is problematic for apps with only 1 cluster because the form presented to the
user does not give an option to change the value. So if you've cached a bad value,
they're stuck with it.