pinned apps to support wildcard globs (#898)
pinned apps to support wildcard globs

refactor Router#pinned_apps to accept the tokens and apps as input and compute
pinned apps. This puts app logic & member variables in the controllers
instead of in Router.

This also adds a OodApp#has_sub_apps? api to determine if an app
in fact has valid sub apps for showing the subapp in the pinned
apps menus.