add_cluster_widget should check for clusters not configured_clusters
add_cluster_widget should check for clusters not configured_clusters
because just because you've configured a cluster does not mean the user
has access to it. Plus all the logic in the method is based off of
clusters, not configured_clusters. This is a bug that can result in
nils being thrown because there are configured_clusters, but no
clusters (i.e., you're not allowed? or you've configured something
that doesn't exist).