change authentication default behaviour
Authentication had previously defaulted to basic auth with folks
updating the .htaccess file. The default is currently Dex, but that
provides a strange state where users defaulted to OIDC but may not
have the OIDC libraries. And on top of that there's a local user
'ood' it defaulted with.

This change means - if there is no auth set, we provide a public
page that directs folks to the documentation for setting up
authenication.

This will force users to make a choice at this step in the installation
process instead of us forcing one on them.