Add the ability to track job state in projects (#2710)
Add the ability to track job state in projects

This adds the ability to track job state in projects. It includes a
new controller and routes for getting the job information. It also
includes a little javascript to update the projects#show page
with the information that it's queried from the jobs controller.