refactor the path selector to accomidate variable paths.
The previous implementation of the path selector had ids hardcoded,
namely working_dir as the field you were trying to select a path for.

This allows for a path selector to be attached to any id. I had to
refactor a lot of it to sort of simplifiy all the event passing
back and forth. It also gets rid of sweet alert because we kinda
want to get rid of it long term anyhow.