work on script initialization and serialization.
There's some wonkiness in Script initialization. namely in adding
require fields like cluster and auto_scripts. This updates all that
to force cluster and auto_scripts to be a part of the script
and serialize correctly. I've added Attribute#options_to_serialize
to help with writing these objects out to yaml as it was filtering
critical options and not writing them to yaml.