Rescues our attempt to find the file by reuploading it
The screenshot from the test failure indicates that the hello-world.c file isn't being uploaded. This failure occurs when I first start up my environment in the morning and then is fixed after I attempt to `binding.pry` into it. I want to see what happens if we attempt to re-upload the file on failure. :person_shrugging: