determin inactive users and clean their PUNs too (#3942)
Determine inactive users and clean their PUNs too. This sends a SIGTERM to the PUN
becuase stopping it through nginx will fail if the user has already been deleted.

It then gathers a list of directories which will be empty soon, and rmdir() them
after sending the SIGTERM signal to all invalid user PUNs. This will result
in the least sleeping, and will avoid using `rm_rf()` as root, which seems
very risky.

---------

Co-authored-by: Simon Westersund <85734631+CSC-swesters@users.noreply.github.com>