fix CI (#4225)
Fix the CI by adding sleep calls all over the place. Not the best thing to do, but it is sort of the only thing I can do at this time to keep moving forward.

Note that all these sleeps are only needed for the github CI - running the CI locally 'just works'.