update dependencies for ruby 3.0+ (#4222)
Update some dependencies for ruby 3.0+. Note this changes the CI a bit to lock bundler across all unit tests for not only simplicity, but also to use bundler features like forcing the ruby platform.

This sets nokogiri to force_ruby_platform to ensure the built .so file works with whatever version of GLIBC we have available. Using the x86-gnu-linux does not work for EL8 and Ubuntu 22.04 because of the version of GLIBC on those distributions is too low.

Note that bundler 2.5.x (currently 2.5.23) is the most recent bundler that still works for Ruby 3.0. So all lock files have been updated to have been bundled with it mostly because the force_ruby_platform flag isn't available across all bundler versions.