Fix unhandled exception with broken symlink in directory (#4735)
* Fix unhandled exception with broken symlink in zip

Dereferencing a symlink with path.realpath raises an exception if it the
symlink is broken, but the exception is never rescued, causing a 0 byte
zip.

* Fix du exit code check

du exists with status 1 if it encounters any error, e.g. a broken
symlink.

Example output seems to have been done without the -s flag.